+++ rebar.config
@@ -16,8 +16,8 @@
  ]}.
 
 {port_specs,
- [{"linux|darwin", "priv/erloci.so", ["c_src/erloci.c"]},
-  {"win32", "priv/erloci.dll", ["c_src/erloci.c"]}
+ [{"linux|darwin", "priv/ocinif.so", ["c_src/ocinif.c"]},
+  {"win32", "priv/ocinif.dll", ["c_src/ocinif.c"]}
  ]}.
 
 {port_env,
