+++ test/connect.config
@@ -3,7 +3,7 @@
 
 #{
 tns => <<"(DESCRIPTION=(ADDRESS_LIST=(ADDRESS="
-         "(PROTOCOL=tcp)(HOST=192.168.1.43)(PORT=1521)))"
+         "(PROTOCOL=tcp)(HOST=127.0.0.1)(PORT=1521)))"
          "(CONNECT_DATA=(SERVICE_NAME=XE)))">>,
 user => <<"scott">>,
 password => <<"regit">>,
